[refactor] StudioContainer 관심사의 분리
- 서브 UI 컴포넌트를 기능 별로 작게 쪼개기
- 서브 컴포넌트에 전달하는 props 줄이기
